fix build seg fault

+++ AnnService/inc/Core/Common/CommonUtils.h
@@ -16,6 +16,7 @@
 #include <string.h>
 #include <vector>
 #include <set>
+#include <cstdint>
 
 #define PREFETCH
 
@@ -33,18 +34,18 @@ namespace SPTAG
             static inline float atomic_float_add(volatile float* ptr, const float operand)
             {
                 union {
-                    volatile long iOld;
+                    volatile int32_t iOld;
                     float fOld;
                 };
                 union {
-                    long iNew;
+                    int32_t iNew;
                     float fNew;
                 };
 
                 while (true) {
-                    iOld = *(volatile long *)ptr;
+                    iOld = *(volatile int32_t *)ptr;
                     fNew = fOld + operand;
-                    if (InterlockedCompareExchange((long *)ptr, iNew, iOld) == iOld) {
+                    if (InterlockedCompareExchange((int32_t *)ptr, iNew, iOld) == iOld) {
                         return fNew;
                     }
                 }
